custom-elements/instantiating-custom-elements/changing-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>Changing IS attribute of the custom element after element is instantiated</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="After a custom element is instantiated, changing the value of the is attribute must not affect this element's custom element type">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(tagName) !== -1) {
            return;
        }
        var name = 'x-' + tagName;
        doc.registerElement(name);
        if (HTML5_TABLE_ELEMENTS.indexOf(tagName) !== -1) {
            doc.body.innerHTML =
                '<table>' +
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>' +
                '</table>';
        } else {
            doc.body.innerHTML =
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>';
        }
        var customElement = doc.querySelector('#custom-element');
        var obj = doc.createElement(tagName);
        assert_class_string(customElement, obj.constructor.name,
            'Custom element must be a ' + obj.constructor.name);
        HTML5_ELEMENTS.forEach(function(value) {
            var name2 = 'x-' + tagName + '-' + value;
            doc.registerElement(name2);
            customElement.setAttribute('is', name2);
            assert_class_string(customElement, obj.constructor.name,
                'Custom element must not be changed and should be a ' +
                obj.constructor.name);
        });
    });
}, 'Changing IS attribute of the custom element after element is instatiated');
</script>
</body>
</html>
